update object get service

+++ src/FSNode/Utils/ByteString.cs
@@ -14,11 +14,11 @@ namespace Neo.FSNode.Utils
             return ByteString.CopyFrom(Helper.Concat(a.ToByteArray(), b.ToByteArray()));
         }
 
-        public static ByteString Range(this ByteString a, int left, int right)
+        public static ByteString Range(this ByteString a, ulong left, ulong right)
         {
             if (a is null)
                 throw new ArgumentNullException(nameof(a));
-            return ByteString.CopyFrom(a.ToByteArray()[left..right]);
+            return ByteString.CopyFrom(a.ToByteArray()[(int)left..(int)right]);
         }
     }
 }
